+++ src/__tests__/app-shell.test.tsx
@@ -4,6 +4,7 @@ import userEvent from "@testing-library/user-event";
 import App from "../App";
 import { ThemeProvider } from "../theme/ThemeProvider";
 import { BoardProvider } from "../board/BoardProvider";
+import { ClipboardProvider } from "../board/ClipboardProvider";
 import { STORAGE_KEYS } from "../constants/storage";
 import { describe, it, expect, beforeEach } from "vitest";
 
@@ -11,7 +12,9 @@ function renderApp() {
   return render(
     <ThemeProvider>
       <BoardProvider>
-        <App />
+        <ClipboardProvider>
+          <App />
+        </ClipboardProvider>
       </BoardProvider>
     </ThemeProvider>,
   );