+++ src/__tests__/card-drag-blur.test.tsx
@@ -3,6 +3,7 @@ import userEvent from "@testing-library/user-event";
 import App from "../App";
 import { ThemeProvider } from "../theme/ThemeProvider";
 import { BoardProvider } from "../board/BoardProvider";
+import { ClipboardProvider } from "../board/ClipboardProvider";
 import { STORAGE_KEYS } from "../constants/storage";
 import { describe, it, expect, beforeEach } from "vitest";
 
@@ -10,7 +11,9 @@ function renderApp() {
   return render(
     <ThemeProvider>
       <BoardProvider>
-        <App />
+        <ClipboardProvider>
+          <App />
+        </ClipboardProvider>
       </BoardProvider>
     </ThemeProvider>,
   );