+++ src/__tests__/view-toggle-integration.test.tsx
@@ -5,13 +5,16 @@ import { describe, it, expect, beforeEach } from "vitest";
 import App from "../App";
 import { ThemeProvider } from "../theme/ThemeProvider";
 import { BoardProvider } from "../board/BoardProvider";
+import { ClipboardProvider } from "../board/ClipboardProvider";
 import { STORAGE_KEYS } from "../constants/storage";
 
 function renderApp() {
   return render(
     <ThemeProvider>
       <BoardProvider>
-        <App />
+        <ClipboardProvider>
+          <App />
+        </ClipboardProvider>
       </BoardProvider>
     </ThemeProvider>,
   );