+++ src/components/__tests__/AnalyticsModal.test.tsx
@@ -4,6 +4,7 @@ import { describe, it, expect, beforeEach } from "vitest";
 import App from "../../App";
 import { ThemeProvider } from "../../theme/ThemeProvider";
 import { BoardProvider } from "../../board/BoardProvider";
+import { ClipboardProvider } from "../../board/ClipboardProvider";
 import { STORAGE_KEYS } from "../../constants/storage";
 import type { Column } from "../../board/types";
 
@@ -11,7 +12,9 @@ function renderApp() {
   return render(
     <ThemeProvider>
       <BoardProvider>
-        <App />
+        <ClipboardProvider>
+          <App />
+        </ClipboardProvider>
       </BoardProvider>
     </ThemeProvider>,
   );