+++ src/main.tsx
@@ -4,12 +4,15 @@ import "./index.css";
 import App from "./App.tsx";
 import { ThemeProvider } from "./theme/ThemeProvider";
 import { BoardProvider } from "./board/BoardProvider";
+import { ClipboardProvider } from "./board/ClipboardProvider";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <ThemeProvider>
       <BoardProvider>
-        <App />
+        <ClipboardProvider>
+          <App />
+        </ClipboardProvider>
       </BoardProvider>
     </ThemeProvider>
   </StrictMode>,